Finally got Net::SCP working.
Main auth problem was not having the public key locally, which
Net::SSH seems to require (see README).  Also had some issues with
scp.upload returning an error when the local file was of type Pathname
rather than String.  And the remote parameter seems to require the
full filename, not just the directory.