You can now CRUD applications via the web console. (COOLINGTOWER-13)
I've also added a dependency on the paperclip gem and added a small
monkey patch to make the gem work properly on JRuby. Multiple patches
for this issue have been sent upstream previously, so we either need
to fork and build our own gem or maintain the monkey patch long term.

The navigation was re-worked to give four main tabs - "dashboard",
"applications", "deployments", and "environments". The "applications"
tab is the only one fully-functioning at the moment.