Create audit_columns library for populating *_by and *_at model columns
Basically, in your model "include AuditColumns" and then when an action
occurs that needs auditing, "audit_action :deployed". The parameter to
audit_action can be any string or symbol. It appends "_at" and "_by" to
the parameter and looks for these attributes on your model. If found,
they are filled in with the current time and current user, respectively.
For now it assumes the current user can be found by a current_user
method in your controllers.

This will get broken out into a gem soon enough but for now it's part
of CT's source.