Add start_failed state for Instances and Environments (STEAM-64)
If an error occurs while starting an instance, both the instance
and its environment will go into a start_failed state. Currently
no indication is given as to why the start failed.

As part of this, I'm removing the EnvironmentWatcher job. It's
more efficient to just update the environment directly after
transitioning instances to running or stopped states.