Initial pass at storing artifacts in cloud (STEAM-36)
Deployable artifacts are now stored in the user's cloud storage.  At
the moment, this means S3 or on the local filesystems depending on
which deltacloud driver being used.

When deploying an artifact, the artifact is first downloaded from the
cloud to SteamCannon and then pushed out to the instances. A future
enhancement will make the instance Agents pull the artifacts from the
cloud storage avoiding that initial copy to SC.

The new storage classes are under app/models/cloud/storage/*, one
class for each supported deltacloud driver. The Cloud::Storage class
presents a paperclip-compatible storage driver so that we can continue
to use paperclip in models. It delegates to the appropriate
cloud-specific storage based on the user's cloud.

All of the storage drivers need more tests written, which I'll write
before closing the JIRA.