Reworked service configure code to defer jboss_as configuration until mod_cluster instance is available.
Since jboss_as needs the mod_cluster instance info before it can be configured, we wait 
for it to become available (prior to this change, mod_cluster tried to trigger configuration 
for the jboss_as instances, which failed depending on the timing of those instances coming up).

It currently assumes that there is at least one (and only one) mod_cluster instance.