Hyperlink instance services (STEAM-124)
Each instance service in the instances list on the environment
show page is now hyperlinked once the service is running if the
service exposes a url (mod_cluster and jboss_as both do).

Part of this change includes converting the instance ajax
refresh from using json properties to sending back a html string
so we don't have to include all the rendering logic in javascript.