Initial stab at viewing service logs (STEAM-126)
This still needs lots of testing, but you should be able
to view any jboss_as or mud_cluster logs by clicking the
View Logs link for the service on the environment show
page.

The log viewing interface still needs some love but right
now it allows tailing or viewing an entire log.