Remove storage_volumes when not needed. [STEAM-133]
We now have a job that watches for storage_volumes that
need to be deleted, and deletes them. We do it in a job
because there is no point in the instance or environment
life cycle where we can be sure that the volume is actually
detached (it still shows as attached for some period even 
when the instance is fully terminated).