When an instance service enters running, update all instances in the cluster with its internal hostame [STEAM-138]
Each instance service gets an internal hostname based on
the service name and the instance number (postgresql01.local,
jboss-as01.local, jboss-as02.local, etc). Each hostname
is distributed to each cluster member and written to /etc/hosts.

The entries are removed when the instance_services are destroyed,
and when a new instance enters the cluster, it triggers updating of
all of the instance service hostnames.