Artifacts now upload to the cloud async (STEAM-230)
We still use paperclip to stage the artifact locally on
SteamCannon while it uploads to the cloud asynchronously. Once
the artifact is uploaded to the cloud, we remove it from local
storage. A user can deploy an artifact before it's uploaded to the
cloud and it will just sit in Pending Deployment until it's
finished uploading.

This is a fairly large change and it needs some additional testing
before marking STEAM-230 as resolved.