First cut at environment event logging [STEAM-98]
This currently logs every state transition for models under (and
including) environment. Each event has an EventSubject, which
is just a proxy object for the real AR record, allowing records
to be deleted w/o affecting the event log.

EventSubjects form a tree using the ancestry gem.

The bulk of the event code is in config/initializers/has_events.rb.
It currently has no tests, and probably should be pulled out and 
turned in to a gem.