Periodically check if unreachable instances are still unreachable (STEAM-113)
We poll on the same schedule as the instance reaper to see if any
unreachable instances are now running or terminated. After a
timeout (currently 48 hours) we give up and assume the instance was
stopped.

Normally when an instance is stopped external to SteamCannon we'll
pick up on that and it will get marked as stopped. The timeout
condition exists so that we don't poll forever if SteamCannon wasn't
able to reach the cloud provider for many hours and the instance was
stopped outside of SteamCannon.